Move enrichMessage helper into its own module
