+++ apps/web/app/(ee)/api/messages/route.ts
@@ -1,7 +1,7 @@
 import { partnerReachableByProgramWhereInput } from "@/lib/api/partners/partner-reachable-by-program-where-input";
 import { getDefaultProgramIdOrThrow } from "@/lib/api/programs/get-default-program-id-or-throw";
 import { withWorkspace } from "@/lib/auth";
-import { enrichMessage } from "@/lib/messages/utils";
+import { enrichMessage } from "@/lib/messages/enrich";
 import {
   PartnerMessagesSchema,
   getPartnerMessagesQuerySchema,
